Add Visual Grid Settings
Add visual grid settings to _variables - This is set to false by default, change to true to and recompile to use the visual grid. Tweak the other variables to taste.

Move grid variable definiitions from _grid.scss to _variables.scss: The $max-width variable needed to be defined in the same file with the visual grid file so that the visual grid width would match the container width, so I just moved all grid variables to variables, because it seemed a better fit.

- Rename _grid.scss to _layout.scss and move from /components to /common, update main.scss accordingly.